requirements.txt
# Credit Tool — Python Dependencies
# Install via: pip install -r requirements.txt
# Activate venv first: F:\credit_tool\venv\Scripts\activate

# --- API Framework ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# --- Data Validation ---
pydantic==2.9.2
pydantic-settings==2.6.1

# --- Database ---
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Task Queue ---
celery==5.4.0
redis==5.2.0

# --- Scheduling ---
celery[redis]==5.4.0

# --- Seed / Fake Data ---
faker==30.8.1
numpy==2.1.3
pandas==2.2.3

# --- ML (M10) ---
scikit-learn==1.5.2
xgboost==2.1.3
joblib==1.4.2

# --- HTTP Client (for llama.cpp HTTP calls + ERP integration) ---
httpx==0.27.2

# --- Utilities ---
python-dotenv==1.0.1
tenacity==9.0.0

# --- Dev / Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2